☀️ 20% – Keep the retype improvements, stop the uncapped run
The return-type sweep was re-run without its 96-halfword cap to see whether far
owners would close too. Thirty-nine rows in, none had, and the ones it was
improving were improving by ones and twos out of hundreds. That is the stop
rule the working method states -- a search with no closes and no new structural
fact is looping -- so it was stopped rather than left to finish.

The improvements it did find are kept: each is a strictly lower residual at an
unchanged span, which is a real step even when it does not close, and it leaves
less for whatever closes the row later. No new bytes.

`make verify` green at 267,374 bytes.

+++ semantic/resource_3ab_c_02001668.c
@@ -50,7 +50,7 @@ extern void Func_02001840(s32 arg0);
 
 extern void Func_080000d0(s32 callback, s32 arg1);
 extern void Func_080091e0(void *record, s32 arg1);
-extern s32 Func_080770c0(s32 flagId);
+extern s32 Func_080770c0(u8 flagId);
 extern void Func_080770c8(s32 flagId);
 extern void Func_080770d0(s32 flagId);
 extern u8 *Func_0808a080(s32 id);
